move eloquent exceptions to exceptions folder and update class appearances accordingly
