+++ tests/Database/DatabaseEloquentModelTest.php
@@ -938,7 +938,7 @@ class DatabaseEloquentModelTest extends TestCase
     }
 
     /**
-     * @expectedException \Illuminate\Database\Eloquent\MassAssignmentException
+     * @expectedException \Illuminate\Database\Eloquent\Exceptions\MassAssignmentException
      * @expectedExceptionMessage name
      */
     public function testGlobalGuarded()
@@ -1598,7 +1598,7 @@ class DatabaseEloquentModelTest extends TestCase
     }
 
     /**
-     * @expectedException \Illuminate\Database\Eloquent\JsonEncodingException
+     * @expectedException \Illuminate\Database\Eloquent\Exceptions\JsonEncodingException
      * @expectedExceptionMessage Unable to encode attribute [objectAttribute] for model [Illuminate\Tests\Database\EloquentModelCastingStub] to JSON: Malformed UTF-8 characters, possibly incorrectly encoded.
      */
     public function testModelAttributeCastingFailsOnUnencodableData()
